Gradio/requirements.txt
openai>=1.0.0
gradio>=4.0.0
pillow>=10.0.0
requests>=2.31.0

